requirements.txt
peewee~=3.13.3
aiogram~=2.8
requests~=2.23.0
aiohttp~=3.6.2
PyYAML~=5.3.1
Pillow~=7.1.2
beautifulsoup4~=4.9.0